\name{coef.rasch}
\alias{coef.rasch}
\title{
  Extract Model Coefficients
}
\description{
  \code{coef} method for class \code{"rasch"}.
}
\usage{
  \method{coef}{rasch}(object, ...)
}

\arguments{
  \item{object}{
    A fitted rasch model.
  }
  \item{...}{
    Further arguments passed to or from other methods.
  }
}