\name{print.rasch}
\alias{print.rasch}
\title{
  Print Rasch Model Fits
}
\description{
  \code{print} method for class \code{"rasch"}.
}
\usage{
  \method{print}{rasch}(x,...)
}

\arguments{
  \item{x}{
    A fitted rasch model.
  }
  \item{...}{
    Further arguments passed to or from other methods.
  }
}